+++ src/common/slurm_protocol_defs.c
@@ -1424,7 +1424,7 @@ extern char *trigger_type(uint32_t trig_type)
 		return "unknown";
 }
 
-extern char *reservation_flags_string(uint16_t flags)
+extern char *reservation_flags_string(uint32_t flags)
 {
 	char *flag_str = xstrdup("");
